Use csgo::StickerId in inventory_changer::inventory::Skin::Sticker struct

+++ Source/InventoryChanger/ItemGenerator/DefaultGenerator.h
@@ -52,7 +52,7 @@ private:
         skin.seed = attributeGenerator.generatePaintKitSeed();
 
         if (Helpers::isMP5LabRats(item.getWeaponID(), gameItemStorage.getPaintKit(item).id))
-            skin.stickers[3].stickerID = static_cast<int>(csgo::StickerId::BlacksiteFoil);
+            skin.stickers[3].stickerID = csgo::StickerId::BlacksiteFoil;
 
         return skin;
     }
